Hide viewer actions
